[FIRRTL] Make CircuitNamespace helper public
Move the `CircuitNamespace` helper struct from the current Grand Central
source files into a public header file. This acts like a symbol table
specifically for `firrtl.circuit` ops and is useful for avoiding name
conflicts when generating additional modules.
